use size3 in Size
