+++ lib/geometry/Size/changeBasis.js
@@ -1,4 +1,4 @@
-const size2 = require('affineplane').size2
+const size3 = require('affineplane').size3
 
 module.exports = function (newBasis) {
   // @Size:changeBasis(newBasis)
@@ -14,7 +14,7 @@ module.exports = function (newBasis) {
   //   a Size
   //
   const pr = this.basis.getTransitionTo(newBasis)
-  const ps = size2.transitFrom(this.size, pr)
+  const ps = size3.transitFrom(this.size, pr)
   const Size = this.constructor
   return new Size(newBasis, ps)
 }
