+++ lib/geometry/Size/getRaw.js
@@ -1,7 +1,7 @@
 module.exports = function () {
   // @Size:getRaw()
   //
-  // Return plain size3 object `{ w, h }` without basis data.
+  // Return plain size3 object `{ w, h, d }` without basis data.
   //
   return this.size
 }
