+++ lib/geometry/Size/index.js
@@ -1,16 +1,25 @@
+const size3 = require('affineplane').size3
+
 const Size = function (basis, size) {
   // @Size(basis, size)
   //
-  // A rectangular size in space. Basically it is two-dimensional distance.
-  // If you need to represent a rectangular shape in space, use Path or Polygon
-  // instead.
+  // A cuboid size in space. Basically it is three-dimensional distance.
+  // If you need to represent a rectangular or cuboid shape in space,
+  // use Path, Polygon, or Box instead as they can be reliably transited
+  // between bases.
   //
   // Parameters
   //   basis
   //     a Basis
   //   size
-  //     a size2 object `{ w, h }`.
+  //     a size3 object `{ w, h, d }`.
   //
+
+  // DEBUG
+  if (!size3.validate(size)) {
+    throw new Error('Invalid size3')
+  }
+
   this.basis = basis
   this.size = size
 }
