+++ lib/geometry/Size/scaleBy.js
@@ -1,4 +1,4 @@
-const size2 = require('affineplane').size2
+const size3 = require('affineplane').size3
 
 module.exports = function (multiplier) {
   // @Size:scaleBy(multiplier)
@@ -12,7 +12,7 @@ module.exports = function (multiplier) {
   // Return
   //   a Size
   //
-  const ms = size2.scaleBy(this.size, multiplier)
+  const ms = size3.scaleBy(this.size, multiplier)
   const Size = this.constructor
   return new Size(this.basis, ms)
 }
