+++ lib/geometry/Size/transitRaw.js
@@ -1,4 +1,4 @@
-const size2 = require('affineplane').size2
+const size3 = require('affineplane').size3
 
 module.exports = function (newBasis) {
   // @Size:transitRaw(newBasis)
@@ -11,8 +11,8 @@ module.exports = function (newBasis) {
   //     a Basis
   //
   // Return
-  //   a size2, an object.
+  //   a size3, an object.
   //
   const tran = this.basis.getTransitionTo(newBasis)
-  return size2.transitFrom(this.size, tran)
+  return size3.transitFrom(this.size, tran)
 }
